The first files I made to start the project
I started with the admin section, in the folder admin you will find
index.php and the application builds starting in that file.  This
project shows how includes, classes, templeting and configs can be used
in an application.  In this particular project one of the goals was
allowing flat file storage or database access. FF was my abv. of Flat
File.